create recommendation

+++ django_project/movies/views.py
@@ -9,6 +9,7 @@ from rest_framework.permissions import IsAuthenticated
 
 from .serializers import *
 from .models import Movie
+import random
 
 from django.contrib.auth import get_user_model
 UserModel = get_user_model()
@@ -25,10 +26,42 @@ def index(request):
 
 # 추천영화 리스트
 @api_view(['GET'])
-def recommendations(request, movie_pk):
-    movie = Movie.objects.get(pk=movie_pk)
-    recommendation_movies = movie.similars.all()
-    serializer = MovieListSerializer(recommendation_movies, many=True)
+def recommendations(request):
+    user = get_object_or_404(UserModel, pk=request.user.id)
+    like_list = user.like_movie.all()
+    like_ids = like_list.values_list('id', flat=True)
+    like_num = len(like_list)
+    random_num = 2
+    
+    recommendation_list = []
+    
+    # 좋아요 한 영화중 3개 선택
+    if like_num >= 3:
+        recommendation_list.extend(random.sample(like_ids, 3))
+    elif 0 < like_num < 3:
+        recommendation_list.extend(random.sample(like_ids, like_num))
+    random_num += (3-like_num)
+    
+    # 나머지 2개 영화(좋아요 영화 장르기반)
+    like_genre = []
+    if like_list:
+        for like in like_list: # 장르 다 합치기
+            like_genre.extend(like.genre)
+    like_genre = list(set(like_genre))
+    
+    if recommendation_list:
+        movies = Movie.objects.exclude(id__in=recommendation_list).distinct()
+    else:
+        movies = Movie.objects.all()
+    
+    if like_genre:
+        choice = movies.filter(genre__in=like_genre).distinct()
+        recommendation_list.extend(random.sample(choice, random_num))
+    else:
+        choice = movies.all().distinct()
+        recommendation_list.extend(random.sample(choice, random_num))
+    
+    serializer = MovieListSerializer(recommendation_list, many=True)
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
@@ -93,14 +126,17 @@ def review_detail(request, review_pk):
         return Response(serializer.data, status=status.HTTP_200_OK)
     
     elif request.method == 'DELETE':
-        review.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.id == review.user_id:
+            review.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        serializer = ReviewSerializer(review, data=request.data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        if request.user.id == review.user_id:
+            serializer = ReviewSerializer(review, data=request.data, partial=True)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(status=status.HTTP_403_FORBIDDEN)
 
 
 # 댓글 작성
@@ -116,16 +152,10 @@ def create_comments(request, review_pk):
 
 
 # 댓글 상세페이지(댓글은 상세조회 없음) - 댓글 수정, 삭제
-@api_view(['PUT', 'DELETE'])
+@api_view(['DELETE'])
 def comment_detail(request, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    
-    if request.method == 'DELETE':
+    if request.user.id == comment.user_id:
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-    elif request.method == 'PUT':
-        serializer = CommentListSerializer(comment, data=request.data, partial=True)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(status=status.HTTP_403_FORBIDDEN)
